all review tag

+++ app/Http/Controllers/Tickets/TicketController.php
@@ -197,19 +197,16 @@ class TicketController extends AbstractController
     public function delete_tag(Request $request) {
         $tag = Tag::find($request->input('tag_id'));
         $ticket = Ticket::find($request->input('ticket_id'));
-
         $tag->taglists()->detach($request->input('taglist_id'));
+
+        $taglist = TagList::find($request->input('taglist_id'));
+        if ($taglist->tags()->count() == 0)
+            $taglist->delete();
         return view('tickets.parts.tags')
             ->with('ticket', $ticket);
     }
 
-    public function delete_ThreadTagList(Request $request) {
-        $taglist = TagList::find($request->input('taglist_id'));
-        $taglist->tags()->detach();
-        $taglist->delete();
-    }
-
-    public function saveThreadTags(Request $request) {
+    public function saveTicketTags(Request $request) {
         $ticket = Ticket::find($request->input('ticket_id'));
         $taglist = TagList::find($request->input('taglist_id'));
         $tag = Tag::findOrFail($request->input('tag_id'));
@@ -223,6 +220,11 @@ class TicketController extends AbstractController
         $taglist->addTag($tag);
         return view('tickets.parts.tags')
             ->with('ticket', $ticket);
+    }
+
+    public function hasTagList(Request $request){
+        $ticket = Ticket::find($request->input('ticket_id'));
+        return $ticket->taglists()->count();
     }
 
     public function clickAndCall(Request $request): JsonResponse
